Refactored taxonomy field xml check and added functionality to ListInstance provisioning and extraction
