+++ Core/OfficeDevPnP.Core/Framework/Provisioning/ObjectHandlers/ObjectField.cs
@@ -184,32 +184,6 @@ namespace OfficeDevPnP.Core.Framework.Provisioning.ObjectHandlers
             return schemaXml;
         }
 
-        private string TokenizeFieldFormula(string fieldXml)
-        {
-            var schemaElement = XElement.Parse(fieldXml);
-            var formula = schemaElement.Descendants("Formula").FirstOrDefault();
-
-            if (formula != null)
-            {
-                var formulaString = formula.Value;
-                if (formulaString != null)
-                {
-                    var fieldRefs = schemaElement.Descendants("FieldRef");
-                    foreach (var fieldRef in fieldRefs)
-                    {
-                        var fieldInternalName = fieldRef.Attribute("Name").Value;
-                        formulaString = formulaString.Replace(fieldInternalName, string.Format("[{{fieldtitle:{0}}}]", fieldInternalName));
-                    }
-                    var fieldRefParent = schemaElement.Descendants("FieldRefs");
-                    fieldRefParent.Remove();
-
-                }
-                formula.Value = formulaString;
-            }
-
-            return schemaElement.ToString();
-        }
-
         private static void CreateField(Web web, XElement templateFieldElement, PnPMonitoredScope scope, TokenParser parser, string originalFieldXml)
         {
             var listIdentifier = templateFieldElement.Attribute("List") != null ? templateFieldElement.Attribute("List").Value : null;
@@ -269,59 +243,7 @@ namespace OfficeDevPnP.Core.Framework.Provisioning.ObjectHandlers
             }
         }
 
-        private static bool IsFieldXmlValid(string fieldXml, TokenParser parser, ClientRuntimeContext context)
-        {
-            var isValid = true;
-            var leftOverTokens = parser.GetLeftOverTokens(fieldXml);
-            if (!leftOverTokens.Any())
-            {
-                var fieldElement = XElement.Parse(fieldXml);
-                if (fieldElement.Attribute("Type").Value == "TaxonomyFieldType")
-                {
-                    var termStoreIdElement = fieldElement.XPathSelectElement("//ArrayOfProperty/Property[Name='SspId']/Value");
-                    if (termStoreIdElement != null)
-                    {
-                        var termStoreId = Guid.Parse(termStoreIdElement.Value);
-                        TaxonomySession taxSession = TaxonomySession.GetTaxonomySession(context);
-                        try
-                        {
-                            taxSession.EnsureProperty(t => t.TermStores);
-                            var store = taxSession.TermStores.GetById(termStoreId);
-                            context.Load(store);
-                            context.ExecuteQueryRetry();
-                            if (store.ServerObjectIsNull.HasValue && !store.ServerObjectIsNull.Value)
-                            {
-                                var termSetIdElement = fieldElement.XPathSelectElement("//ArrayOfProperty/Property[Name='TermSetId']/Value");
-                                if (termSetIdElement != null)
-                                {
-                                    var termSetId = Guid.Parse(termSetIdElement.Value);
-                                    try
-                                    {
-                                        var termSet = store.GetTermSet(termSetId);
-                                        context.Load(termSet);
-                                        context.ExecuteQueryRetry();
-                                        isValid = termSet.ServerObjectIsNull.HasValue && !termSet.ServerObjectIsNull.Value;
-                                    }
-                                    catch (Exception)
-                                    {
-                                        isValid = false;
-                                    }
-                                }
-                            }
-                        }
-                        catch (Exception)
-                        {
-                            isValid = false;
-                        }
-                    }
-                    else
-                    {
-                        isValid = false;
-                    }
-                }
-            }
-            return isValid;
-        }
+       
 
         private static void ValidateTaxonomyFieldDefaultValue(TaxonomyField field)
         {
